requirements.txt
django
python-decouple
psycopg2-binary
djangorestframework
black
